requirements.txt
django
djangorestframework
psycopg2-binary
vaderSentiment
django-cors-headers
